call mkstemp with a nondir prefix
Older versions of python take prefix literally. This becomes a problem
if file is not in CWD, as the attempted tempfile will contain duplicated
segments of dirname. For example given sub/dir/foo.txt the tempfile
attempted to be created will be sub/dir/sub/dir/foo.txt.XXXXX, which
probably does not exist.

closes #7